backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pywebview==4.4.1
sqlalchemy==2.0.29
pyinstaller>=6.5.0
